+++ sql/llx_document_model.sql
@@ -14,7 +14,7 @@
 -- along with this program.  If not, see https://www.gnu.org/licenses/.
 
 INSERT INTO llx_document_model (nom, entity, type, libelle, description) VALUES ('standard_equipment', '1', 'equipment', NULL, NULL);
-INSERT INTO llx_document_model (nom, entity, type, libelle, description) VALUES ('standard_equipmentlist', '1', 'equipment', NULL, NULL);
+INSERT INTO llx_document_model (nom, entity, type, libelle, description) VALUES ('lims_equipmentlist', '1', 'equipment', NULL, NULL);
 INSERT INTO llx_document_model (nom, entity, type, libelle, description) VALUES ('lims_testreport', '1', 'samples', NULL, NULL);
 INSERT INTO llx_document_model (nom, entity, type, libelle, description) VALUES ('standard_methods', '1', 'methods', NULL, NULL);
 INSERT INTO llx_document_model (nom, entity, type, libelle, description) VALUES ('standard_limits', '1', 'limits', NULL, NULL);